% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_colormeter.R
\name{legend-coords}
\alias{legend-coords}
\title{Colormeter legend coordinate system}
\description{
Due to the peculiar nature of the colormeter legend (e.g., dynamic size and positioning
of color bars/arcs depending on the number of breaks), many parts of the legend are drawn in a
mini isolated coordinate space. You can inspect this coordinate space with \code{debug = TRUE}.

The coordinate space defaults to the x-y range of the legend "data", which considers the
polar-transformed positioning of the color arcs/bars, the labels, and the dashboard circle.
}
\keyword{internal}
